Update Webscraping.py
I added the URLI think I want to use and I made a change to my comment on line 69. 

OG: # Modulus to cycle through words

New: # Modulus to cycle through words based on the day

I also want to add in Twilio. Right now, this code is "set up" to try and email this to you every day, but I RARELY check my email unless I know something is coming so I want to send it to "the users' email", but I also want to send it to the user via a text message or a notification. I found Twilio